upgrade keyless example

+++ examples/keyless-example/src/core/ephemeral.ts
@@ -1,7 +1,8 @@
 // Copyright © Aptos
 // SPDX-License-Identifier: Apache-2.0
 
-import { Ed25519PrivateKey, EphemeralKeyPair } from "@aptos-labs/ts-sdk";
+import { Ed25519PrivateKey } from "@aptos-labs/ts-sdk";
+import { EphemeralKeyPair } from "@aptos-labs/ts-sdk/keyless";
 
 export const EphemeralKeyPairEncoding = {
   decode: (e: any) => EphemeralKeyPair.fromBytes(e.data),
@@ -27,7 +28,7 @@ export const isValidEphemeralKeyPair = (keyPair: EphemeralKeyPair): boolean => {
  * @param params Additional parameters for the ephemeral key pair
  */
 export const createEphemeralKeyPair = ({
-  expiryDateSecs = BigInt(Math.floor(Date.now() / 1000)) + BigInt(24 * 60 * 60),
+  expiryDateSecs = Math.floor(Date.now() / 1000) + 24 * 60 * 60,
   privateKey = Ed25519PrivateKey.generate(),
   ...options
 }: Partial<ConstructorParameters<typeof EphemeralKeyPair>[0]> = {}) =>
